+++ CHANGELOG.md
@@ -1,10 +1,17 @@
 #### 0.3.8 (2018-12-03)
 
+##### Documentation Changes
+
+*  Build docs for comps ([c0ca726e](https://github.com/big-neon/bn-api-node/commit/c0ca726e0eadf040b5f2594964d2d23ddafa2a87))
+
 ##### New Features
 
 * **endpoints:**
+  *  comps.[update, del, read] endpoints ([1c3e8a5d](https://github.com/big-neon/bn-api-node/commit/1c3e8a5d35b9888a25960073f2ab27e16d0b9d79))
+  *  Added /events/checkins endpoint ([#238](https://github.com/big-neon/bn-api-node/pull/238)) ([58017291](https://github.com/big-neon/bn-api-node/commit/5801729119523fab9404045ade99b2f30edcdbe4))
   *  Added organizations.fans.[read, history, index] ([0c4f35f7](https://github.com/big-neon/bn-api-node/commit/0c4f35f7b0307127d4ac3f3681495f6198d65173))
   *  Added /events/checkins endpoint ([#238](https://github.com/big-neon/bn-api-node/pull/238)) ([58017291](https://github.com/big-neon/bn-api-node/commit/5801729119523fab9404045ade99b2f30edcdbe4))
+
 
 #### 0.3.6 (2018-11-29)